+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.11.0 
+- Multiple story arc in single issue title
+- Option to strip lead "The " from story arc. Default False
+- Option to end Story Arc on colon character. Default True
+
 v0.10.0
 - Config menu option to set field to save found story arc into either "Story Arc" or "Alternate Series"